他ユーザーからの編集機能へのアクセス制限まで

+++ app/controllers/books_controller.rb
@@ -9,8 +9,14 @@ class BooksController < ApplicationController
   def create
     @booknew = Book.new(book_params)
     @booknew.user_id = current_user.id
-    @booknew.save
-    redirect_to book_path(@booknew[:id])
+    if @booknew.save
+      flash[:notice] = "You have created book successfully."
+      redirect_to book_path(@booknew[:id])
+    else
+      @user = current_user
+      @books = Book.all
+      render :index
+    end
   end
 
   def show
@@ -21,20 +27,34 @@ class BooksController < ApplicationController
   end
 
   def edit
+    book = Book.find(params[:id])
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
     @book = Book.find(params[:id])
   end
 
   def update
+    book = Book.find(params[:id])
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
+
     @book = Book.find(params[:id])
       if  @book.update(book_params)
-        flash[:notice] = "Book was successfully updated."
+        flash[:notice] = "You have updated book successfully."
         redirect_to book_path(@book.id)
      else
-        render :edit
+     render :edit
      end
-   end
+  end
 
   def destroy
+    book = Book.find(params[:id])
+    unless book.user.id == current_user.id
+      redirect_to books_path
+    end
+
     book = Book.find(params[:id])
     book.destroy
     redirect_to books_path
